+++ assets/package.json
@@ -1,15 +1,16 @@
 {
   "dependencies": {
     "bulma": "^0.8.0",
-    "css-loader": "^3.4.2",
     "flatpickr": "^4.6.3",
+    "vue": "^2.6.11"
+  },
+  "devDependencies": {
+    "css-loader": "^3.4.2",
+    "expose-loader": "^0.7.5",
     "sass-loader": "^8.0.2",
     "style-loader": "^1.1.3",
     "ts-loader": "^6.2.1",
     "typescript": "^3.7.5",
-    "vue": "^2.6.11"
-  },
-  "devDependencies": {
     "webpack": "^4.41.5",
     "webpack-cli": "^3.3.10"
   }
