+++ assets/src/entry.ts
@@ -1,0 +1,13 @@
+import Vue from 'vue';
+import {DjetlerApp} from "./DjetlerApp";
+
+let djetlerApp = new DjetlerApp();
+
+
+let djetlerVueTest = new Vue({
+    el: "#djetler-vue",
+    delimiters: ["[[", "]]"],
+    data: {
+        message: "Hello Djetler!!!"
+    }
+});
