+++ django_ledger/templates/django_ledger/base.html
@@ -36,6 +36,7 @@
     </script>
     <script src="{% static 'django_ledger/js/flatpickr.min.js' %}"></script>
     <script src="{% static 'django_ledger/js/djetler_scripts.js' %}"></script>
+    <script src="{% static 'django_ledger/dist/djetler.bundle.js' %}"></script>
 {% endblock %}
 
 </body>
