+++ django_ledger/templates/django_ledger/entity_detail.html
@@ -32,7 +32,9 @@
                             <article class="tile is-child box content">
                                 <h2 class="has-text-centered has-text-weight-light">{% trans 'Insights' %}</h2>
                                 <p>Future place for cool stuff.</p>
+
                                 <div id="djetler-vue">[[message]]</div>
+
                             </article>
                         </div>
                     </div>
